requirements.txt
discord.py
requests
bs4
beautifulsoup4
imap_tools